feat: add resposive design
