+++ client/components/Card/Card.scss
@@ -5,6 +5,14 @@
   width: 300px;
   margin-bottom: 50px;
 
+  @media(max-width: 960px) {
+    width: 250px;
+}
+
+@media(max-width: 768px) {
+  width: 300px;
+}
+
   & > button {
     @include roundLink;
     @include activeLink;
