+++ client/components/Catalog/Catalog.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import Category from '../Category'
+import Cards from '../Cards'
+
+const Main = () => {
+  return (
+    <main>
+      <Category text="pizza" />
+      <Cards />
+    </main>
+  )
+}
+
+export default Main
