+++ client/components/Category/Category.scss
@@ -1,4 +1,7 @@
+@import '../../styles/variables';
+
 .category {
+    @include fontHeader;
     font-size: 40px;
     margin-bottom: 40px;
     text-transform: uppercase;
