+++ client/components/Filter/Filter.scss
@@ -1,0 +1,18 @@
+@import '../../styles/variables';
+
+.filter {
+    &__title {
+        margin-bottom: 20px;
+        color: $color-text-text;
+        font-size: 16px;
+        text-transform: uppercase;
+    }
+
+    &__settings {
+        color: $color-text-text;
+        font-size: 18px;
+        text-transform: uppercase;
+        font-weight: bold;
+        letter-spacing: 1px;
+    }
+}
