+++ client/components/Filter/HashTag/Hashtag.scss
@@ -3,6 +3,8 @@
 .hashtag {
     text-transform: uppercase;
     color: $color-text-header;
-    font-size: 14px;
+    font-size: 16px;
     margin-right: 10px;
+    letter-spacing: 0.5px;
+    font-weight: bold;
 }
