--- client/components/Main/index.jsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import Category from '../Category'
-import Cards from '../Cards'
-
-const Main = () => {
-  return (
-    <main>
-      <Category text="pizza" />
-      <Cards />
-    </main>
-  )
-}
-
-export default Main
