+++ client/components/SideBar/sideBar.scss
@@ -4,21 +4,62 @@
   @include fixed(fixed);
   @include colomnFlex(center, flex-start);
   flex-direction: column;
-  width: $sidebar-width-768;
+  width: $sidebar-width-1024;
   background-color: $color-sideBar;
   padding-left: 35px;
 
+  @media (max-width: 1024px) {
+    width: $sidebar-width-768;
+}
+
+@media (max-width: 768px) {
+  position: relative;
+  width: 100%;
+}
+
+nav {
+  width: 100%;
+}
+
   & > span {
     opacity: 0.6;
+    text-transform: uppercase;
   }
 
-  a {
+  &__nav {
     @include colomnFlex(flex-end, flex-start);
     width: 100%;
+  }
+
+  &__nav-link {
     height: 50px;
+    width: 100%;
+
+    a {
+      @include colomnFlex(flex-end, flex-start);
+    @include equalSides(100%);
     font-size: 20px;
     color: $color-text-header;
     font-weight: 600;
     text-transform: uppercase;
+    }
   }
+
+  &__nav-link.active {
+    position: relative;
+    padding-left: 30px;
+
+    &:before {
+      content: "";
+      display: block;
+      width: 20px;
+      height: 1px;
+      background: #000;
+      left: 0;
+      top: 75%;
+      position: absolute;
+    }
+  }
+
+  
 }
