--- client/pages/main/mainPage.scss
@@ -1,13 +1,0 @@
-@import '../../styles/variables';
-
-.page_main {
-  display: flex;
-  justify-content: flex-end;
-  min-height: 100vh;
-  max-height: 100%;
-
-  main {
-    z-index: 10;
-    width: 105%;
-  }
-}
